moved enroller analytics files around and cleaned up excessive styling
